backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4==4.13.4",
    "celery==5.5.3",
    "django==5.2",
    "django-celery-beat==2.8.1",
    "django-cors-headers==4.7.0",
    "django-environ==0.12.0",
    "django-filter==25.1",
    "djangorestframework==3.16.0",
    "djangorestframework-simplejwt==5.5.0",
    "drf-standardized-errors==0.14.1",
    "fake-useragent==2.2.0",
    "flower==2.0.1",
    "gunicorn==25.1.0",
    "openai==2.13.0",
    "pillow==12.0.0",
    "psycopg[binary,pool]==3.2.9",
    "pydantic==2.12.3",
    "redis==6.1.0",
    "requests==2.32.4",
    "whitenoise==6.11.0",
]

[dependency-groups]
dev = [
    "django-extensions==4.1",
    "django-silk==5.4.0",
    "django-stubs==5.2.0",
    "djangorestframework-stubs==3.16.0",
    "flake8==7.2.0",
    "flake8-bugbear==24.12.12",
    "flake8-comprehensions==3.16.0",
    "mypy==1.15.0",
    "types-requests==2.32.4.20250611",
    "werkzeug==3.1.3",
]
